Moved common types to separate file
